move elementwise operators code to separate file
